Split attestation functions from pal into separate header
